add experimental code that reads relations and writes to a Berkeley database

+++ branches/msieve-db/gnfs/filter/filter.h
@@ -31,8 +31,8 @@ extern "C" {
    value is the large prime bound to use for the singleton removal */
 
 uint32 nfs_purge_duplicates(msieve_obj *obj, factor_base_t *fb,
-				uint32 max_relations,
-				uint32 *num_relations_out); 
+				uint64 max_relations,
+				uint64 *num_relations_out); 
 
 /* read '<savefile_name>.d' and create '<savefile_name>.lp', a 
    binary file containing the relations surviving the singleton
